[Refactor] Rename project
