+++ eupi-pdf-controller/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.euphony.eupi_ppt_controller">
+    package="com.euphony.eupi_pdf_controller">
 
     <uses-permission android:name="android.permission.RECORD_AUDIO" />
