+++ eupi-pdf-controller/src/main/java/com/euphony/eupi_pdf_controller/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.euphony.eupi_ppt_controller.ui.theme
+package com.euphony.eupi_pdf_controller.ui.theme
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material.MaterialTheme
