+++ eupi-pdf-controller/src/test/java/com/euphony/eupi_pdf_controller/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.euphony.eupi_ppt_controller
+package com.euphony.eupi_pdf_controller
 
 import org.junit.Test
