+++ eupi-pdf-viewer/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 33
 
     defaultConfig {
-        applicationId "com.euphony.eupi_ppt_viewer"
+        applicationId "com.euphony.eupi_pdf_viewer"
         minSdk 26
         targetSdk 29
         versionCode 1
