+++ eupi-pdf-viewer/src/androidTest/java/com/euphony/eupi_pdf_viewer/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.euphony.eupi_ppt_controller
+package com.euphony.eupi_pdf_viewer
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.euphony.eupi_ppt_controller", appContext.packageName)
+        assertEquals("com.euphony.eupi_pdf_viewer", appContext.packageName)
     }
 }
