+++ eupi-pdf-viewer/src/main/java/com/euphony/eupi_pdf_viewer/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.euphony.eupi_ppt_viewer
+package com.euphony.eupi_pdf_viewer
 
 import android.content.Context
 import android.content.Intent
