+++ eupi-pdf-viewer/src/test/java/com/euphony/eupi_pdf_viewer/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.euphony.eupi_ppt_viewer
+package com.euphony.eupi_pdf_viewer
 
 import org.junit.Test
